+++ src/main/web/report/finalist/non-numeric_ui.js
@@ -178,14 +178,6 @@ const nonNumericUi = {}
         return "org_" + category + "_" + teamIdx;
     }
 
-    function teamJudgingStationId(category, teamIdx) {
-        return "judgingStation_" + category + "_" + teamIdx;
-    }
-
-    function teamJudgesId(category, teamIdx) {
-        return "judges_" + category + "_" + teamIdx;
-    }
-
     function teamDeleteId(category, teamIdx) {
         return "delete_" + category + "_" + teamIdx;
     }
@@ -200,18 +192,6 @@ const nonNumericUi = {}
 
         const teamOrgElement = document.getElementById(teamOrgId(category.catId, teamIdx));
         teamOrgElement.value = team.org;
-
-        const teamJudgingStationElement = document.getElementById(teamJudgingStationId(category.catId, teamIdx));
-        teamJudgingStationElement.value = team.judgingGroup;
-
-        const judges = finalist_module.getNominatingJudges(category, team.num);
-        let judgesStr;
-        if (!judges) {
-            judgesStr = "";
-        } else {
-            judgesStr = judges.filter(x => x).join(", ");
-        }
-        document.getElementById(teamJudgesId(category.catId, teamIdx)).value = judgesStr;
     }
 
     /**
@@ -239,7 +219,6 @@ const nonNumericUi = {}
                 finalist_module.removeTeamFromCategory(category, prevTeam);
                 document.getElementById(teamNameId(category.catId, teamIdx)).value = "";
                 document.getElementById(teamOrgId(category.catId, teamIdx)).value = "";
-                document.getElementById(teamJudgingStationId(category.catId, teamIdx)).value = "";
                 if (_useStorage) {
                     finalist_module.saveToLocalStorage();
                 }
@@ -276,18 +255,6 @@ const nonNumericUi = {}
         orgEle.setAttribute("id", teamOrgId(category.catId, teamIdx));
         orgEle.readonly = true;
         orgEle.disabled = true;
-
-        const judgingStationEle = document.createElement("input");
-        teamEle.appendChild(judgingStationEle);
-        judgingStationEle.setAttribute("id", teamJudgingStationId(category.catId, teamIdx));
-        judgingStationEle.readonly = true;
-        judgingStationEle.disabled = true;
-
-        const judgesEle = document.createElement("input");
-        teamEle.appendChild(judgesEle);
-        judgesEle.setAttribute("id", teamJudgesId(category.catId, teamIdx));
-        judgesEle.readonly = true;
-        judgesEle.disabled = true;
 
         const deleteButton = document.createElement("button");
         teamEle.appendChild(deleteButton);